[fix] entity 및 mapper 단순 코드 정리
